Refactor:  코드 리팩토링 및 파일 분리, 컴포넌트  파일 확장자 jsx로 변경
1. 카드 생성 날짜 및 경과시간 계산 및 표시 기능 코드 리팩토링 및 CardList컴포넌트에서 파일 분리 후 utils 폴더로 이동
2. SNS 데이터 Footer컴포넌트에서 분리 후 data 폴더로 이동
3. 컴포넌트 파일 확장자 js에서 jsx로 변경
4. useAsync 커스텀 훅 구현 후  App 컴포넌트에 사용
