+++ src/hooks/useAsync.js
@@ -1,0 +1,24 @@
+import { useCallback, useState } from "react";
+
+function useAsync(asyncFunction) {
+  const [pending, setPending] = useState(false);
+  const [error, setError] = useState(null);
+
+  const wrappedFunction = useCallback(
+    async (...args) => {
+      setPending(true);
+      setError(null);
+      try {
+        return await asyncFunction(...args);
+      } catch (error) {
+        setError(error);
+      } finally {
+        setPending(false);
+      }
+    },
+    [asyncFunction]
+  );
+
+  return [pending, error, wrappedFunction];
+}
+export default useAsync;
