+++ src/services/api.js
@@ -11,8 +11,7 @@ const API_URL = "https://bootcamp-api.codeit.kr/api";
 //   return response;
 // }
 export async function getApiItems(apiPath) {
-  const path = `${apiPath}`;
-  const response = await fetch(`${API_URL}/${path}`);
+  const response = await fetch(`${API_URL}/${apiPath}`);
   if (!response.ok) {
     throw new Error("데이터를 불러오는데 실패했습니다");
   }
